requirements.txt
# Audio Processing Toolkit - Complete Requirements
# Supports complete 8-step workflow: cleanup → extraction → segmentation → transcription → samples → translation → TTS → merging → vocal removal

# Core audio/video processing
moviepy>=1.0.3
librosa>=0.9.0
soundfile>=0.10.0

# Deep learning and GPU acceleration
torch>=1.9.0
torchaudio>=0.9.0

# Speaker diarization and audio analysis
pyannote.audio>=3.1.0

# Speech recognition and transcription
openai-whisper>=20231117

# Scientific computing and signal processing
numpy>=1.21.0
scipy>=1.7.0

# Translation dependencies (NLLB-200)
transformers>=4.20.0
sentencepiece>=0.1.96

# TTS dependencies for voice cloning (XTTS-v2)
TTS>=0.22.0

# Advanced tokenizer for XTTS
tokenizers>=0.13.0

# Audio processing utilities
pydub>=0.25.1
webrtcvad>=2.0.10

# File and path handling
pathlib2>=2.3.0; python_version < "3.4"

# JSON and data processing
jsonschema>=3.2.0

# Progress bars and user interface
tqdm>=4.64.0

# Date and time handling
python-dateutil>=2.8.0

# Computer vision and image processing (for face detection)
opencv-python>=4.5.0
matplotlib>=3.5.0

# Vocal removal (Spleeter)
spleeter>=2.3.0

# Optional: GPU monitoring (for advanced users)
# gpustat>=1.0.0 